move nav search into src, move css into each component.  Refacor main scss and pivot into sub files
