--- app/assets/javascripts/angular/components/nav-search/nav-search.scss
@@ -1,8 +1,0 @@
-nav-search {
-    .w-18-rem {
-        width: 18rem;
-    }
-    .mxhi-20-rem {
-        max-height: 20rem;
-    }
-}
