+++ app/assets/stylesheets/main.scss
@@ -1,11 +1,16 @@
 @import './bootstrap';
 @import './variables/colors';
+@import './components/modal';
+@import './components/forms';
 @import './variables/generateColors';
-@import './components/global';
-@import './pivot_theme';
+//@import './components/global';
+
+@import './components/app';
+
+@import './components/buttons';
+@import './components/timepicker';
 @import './components/dropdown';
 @import './components/grid';
-@import './components/header';
 @import './components/navigation';
 @import './components/row';
 @import './components/sheet';
