+++ index.ejs
@@ -22,7 +22,7 @@
 
     <app
         current-organization="8"
-        access="true"    jwt-token="eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoyMzk1MDIxLCJleHAiOjE1NDExMTI3ODh9.X4WPia4ok0ykr6LdTqoh9MUEGDKsoJBcMn6YDdu4hfs"
+        access="true"    jwt-token="eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoyMzk1MDIxLCJleHAiOjE1NDE1MjQzNzZ9.hIE1v_1ILLzVLZj6uXj904QdLLOTolPx5I_Rerp2wzU"
         branded="true"
       >
       </app>
