+++ src/app.scss
@@ -1,0 +1,40 @@
+app {
+    display: flex;
+    min-height: 100vh;
+    flex-direction: column;
+    justify-content: space-between;
+
+    > header {
+        margin-bottom: 140px;
+    }
+
+    .sub-navigation {
+        height: 53px;
+        position: fixed;
+        top: 64px;
+        left: 0;
+        right: 0;
+        z-index: 50;
+        box-shadow: 0 0 10px #808284;
+    }
+
+    .main-container {
+        margin-right: auto;
+        margin-left: auto;
+        padding-left: 15px;
+        padding-right: 15px;
+        min-width: 765px; //Currently below this width components mess up so locking here for now.
+
+        @media screen and (min-width: 768px) {
+            width: 765px;
+        }
+
+        @media screen and (min-width: 992px) {
+            width: 976px;
+        }
+
+        @media screen and (min-width: 1352px) {
+            width: 976px;
+        }
+    }
+}
