+++ src/components/login.scss
@@ -1,0 +1,31 @@
+.btn-key {
+    &:before {
+        background: url(/app/assets/images/icons/logo-key-1x.png) no-repeat
+            center center;
+        width: 20px;
+        height: 22px;
+        display: block;
+        content: '';
+        margin-right: 19px;
+        position: relative;
+        top: 2px;
+        left: 0;
+        float: left;
+    }
+}
+
+.btn-facebook {
+    &:before {
+        background: url(/app/assets/images/icons/facebook-1x.png) no-repeat
+            center center;
+        width: 20px;
+        height: 22px;
+        display: block;
+        content: '';
+        margin-right: 19px;
+        position: relative;
+        top: 2px;
+        left: 0;
+        float: left;
+    }
+}
