+++ src/components/navigation/navHeader.component.js
@@ -1,4 +1,6 @@
 import template from './navHeader.html';
+import './navHeader.scss';
+import './navSearch.component';
 
 angular.module('missionhubApp').component('navHeader', {
     controller: navHeaderController,
