+++ src/components/navigation/navSearch.html
@@ -1,4 +1,4 @@
-<div id="">
+<div>
     <input type="text"
            class="search-autocomplete-field pivot-white"
            placeholder="{{ 'people.search.search_placeholder' | t }}"
